# Implement a few performance tests