--- a/package.json
+++ b/package.json
@@ -14,6 +14,12 @@
         ],
         "outputs": []
       },
+      "test:perf": {
+        "dependsOn": [
+          "^build"
+        ],
+        "outputs": []
+      },
       "lint": {
         "outputs": []
       },
@@ -28,7 +34,7 @@
   ],
   "scripts": {
     "clean": "yarn workspaces foreach run clean",
-    "build": "npx turbo run lint build test"
+    "build": "npx turbo run lint build test test:perf"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^4.28.1",