--- a/packages/arbor-store/package.json
+++ b/packages/arbor-store/package.json
@@ -24,6 +24,7 @@
     "url": "https://github.com/drborges"
   },
   "main": "./dist/index.cjs.js",
+  "type": "module",
   "module": "./dist/index.esm.js",
   "types": "./dist/index.d.ts",
   "license": "MIT",
@@ -48,6 +49,7 @@
     "clean": "rm -rf dist",
     "lint": "yarn eslint . --ext .js,.jsx,.ts,.tsx",
     "test": "jest --passWithNoTests",
+    "test:perf": "node --stack-size=200000 --expose-gc perf/index.mjs",
     "dev": "jest --watch",
     "build": "yarn clean && node ../../tools/build.js && yarn tsc",
     "prepublish": "yarn build",