--- a/packages/arbor-store/package.json
+++ b/packages/arbor-store/package.json
@@ -24,7 +24,6 @@
     "url": "https://github.com/drborges"
   },
   "main": "./dist/index.cjs.js",
-  "type": "module",
   "module": "./dist/index.esm.js",
   "types": "./dist/index.d.ts",
   "license": "MIT",