--- a/packages/arbor-store/perf/helpers.mjs
+++ b/packages/arbor-store/perf/helpers.mjs
@@ -1,4 +1,4 @@
-const defaultDepth = 10000
+export const defaultDepth = 10000
 
 export const createDeeplyNestedState = (depth = defaultDepth, state = {}) => {
   const level = `level${depth}`