--- a/packages/arbor-store/perf/index.mjs
+++ b/packages/arbor-store/perf/index.mjs
@@ -1,0 +1,21 @@
+import { prettyPringMeasurements } from "./helpers.mjs"
+import { checkMemoryAllocationWhenProduceNextStateViaArbor } from "./memory.perf.mjs"
+import {
+  produceNextStateViaArbor,
+  produceNextStateViaManualReducer,
+} from "./mutations.perf.mjs"
+
+import {
+  readingLeafNodeFromArborStateTree,
+  readingLeafNodeFromPlainObjectStateTree,
+} from "./traversing.perf.mjs"
+
+readingLeafNodeFromArborStateTree()
+readingLeafNodeFromPlainObjectStateTree()
+
+produceNextStateViaArbor()
+produceNextStateViaManualReducer()
+
+prettyPringMeasurements(performance.getEntriesByType("measure"))
+
+checkMemoryAllocationWhenProduceNextStateViaArbor()