--- a/packages/arbor-store/perf/index.mjs
+++ b/packages/arbor-store/perf/index.mjs
@@ -2,6 +2,8 @@ import { prettyPringMeasurements } from "./helpers.mjs"
 import { checkMemoryAllocationWhenProduceNextStateViaArbor } from "./memory.perf.mjs"
 import {
   produceNextStateViaArbor,
+  produceNextStateViaArborProducer,
+  produceNextStateViaImmerProducer,
   produceNextStateViaManualReducer,
 } from "./mutations.perf.mjs"
 
@@ -14,6 +16,8 @@ readingLeafNodeFromArborStateTree()
 readingLeafNodeFromPlainObjectStateTree()
 
 produceNextStateViaArbor()
+produceNextStateViaArborProducer()
+produceNextStateViaImmerProducer()
 produceNextStateViaManualReducer()
 
 prettyPringMeasurements(performance.getEntriesByType("measure"))