--- a/packages/arbor-store/perf/mutations.perf.mjs
+++ b/packages/arbor-store/perf/mutations.perf.mjs
@@ -1,6 +1,6 @@
 import { performance } from "perf_hooks"
 
-import Arbor from "../dist/index.esm.js"
+import Arbor from "../dist/index.mjs"
 import {
   createDeeplyNestedState,
   readRecursevly,