Initial outcome:

```js
Reading a leaf node within a very tall Arbor state tree
  Duration: 0.12387502193450928 ms
Reading a leaf node within a very tall but cached Arbor state tree
  Duration: 0.0038330554962158203 ms
Reading a leaf node within a very tall plain object state tree
  Duration: 0.01333308219909668 ms
Producing next state tree via Arbor mutation (depth = 10000)
  Duration: 0.14329099655151367 ms
Producing next state tree via Arbor producer (depth = 10000)
  Duration: 0.03345799446105957 ms
Producing next state tree via Immer producer (depth = 10000)
  Duration: 0.8182079792022705 ms
Producing next state tree via plain reducer (depth = 10000)
  Duration: 0.5309170484542847 ms

Heap Allocation
  After raw state allocation 3198.504 Kb
  After Arbor store allocation 11.408 Kb
  After accessing leaf node 2.088 Kb
  After mutating leaf node 111.36 Kb
```